+++ src/main/java/com/rocketmiles/raygun/KeyMaster.java
@@ -21,7 +21,7 @@
  *  THE SOFTWARE.
  */
 
-package com.fatboyindustrial.raygun;
+package com.rocketmiles.raygun;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Optional;